fix: thread rerender issue
